some update (int number1 div int number2 ==> 0) : data.BatteryPercent = int((float64(bat[0].CurrentCapacity) / float64(bat[0].MaxCapacity)) * 100) read battery information from remote server ./grofer --serverip 127.0.0.1
1)grpc interface
2)ssh remote exec

+++ pkg/metrics/general/serve_stats_darwin.go
@@ -111,7 +111,7 @@ func ServeBattery(ctx context.Context, cpuChannel chan AggregatedMetrics, server
 	if err != nil {
 		return err
 	} else {
-		data.BatteryPercent = int((float32(bat[0].CurrentCapacity) / float32(bat[0].MaxCapacity)) * 100)
+		data.BatteryPercent = int((float64(bat[0].CurrentCapacity) / float64(bat[0].MaxCapacity)) * 100)
 	}
 
 	select {
